Initial Client and Server Implementation
The client and server have basic functionality. The server supports
multiple concurrent clients connecting and sending messages to it. The
client can connect to the server, send messages to it, and receive
messages from the server. Upon receipt of a new message, the server
broadcasts the messages to every client in the same order it received
the messages.

The client's output is very buggy and needs to be fixed.

The server needs to enforce unique usernames per client and send the
username of the sender with every broadcast message. The server also
needs to be able to handle clients leaving without crashing (as it
crashes when that happens now).