Add GUI to client
Make it so the client has a GUI that displays incoming chat messages
received from the server. Currently, only the chat messages received
from the server are shown in the GUI. Next steps would be to add an
input text box to the GUI and a send button to send messages to the chat
room. As of this commit, the chat GUI will receive new messages from the
server, display them to the text box, and auto scroll to the newest
message.