Add input message box and send button to GUI
Added an input message box to the bottom of the GUI where the user can
input the message they'd like to send. Also, added a send button to send
the message to the server. Improved geometry of window and widget
placement. Added send icon image. The image was found on the internet
under a creative commons license that does not require attribution.
Finally, made it so the users cannot edit the chat box that shows
incoming messages from the server.