Reduce CPU consumption per server thread
The server was consuming up to 23% of my CPU during testing on my
computer when just 1 client joined. I added a 1ms delay to the send and
receive thread loops. This dropped the CPU consumption of the server to
around 1% with 2 clients connected. If we find the server's performance
is poor, we can decrease the delay, but I'd say this is good until
proven otherwise.