Bugfix: Make client disconnect gracefully
The disconnect was buggy and the socket was not properly getting closed
on the client and server end. Move the socket close call to the function
that executes when the GUI is closed. Also, make socket operations have
a timeout so we don't block indefinitely on receive calls.

Remove the ctrl-c catching code because it did not work. We will likely
need to set up some sort of event stream to communicate across threads.